Made fixes to data still and the manager
In data still the new filter FilterAdd was created which works more in
the same way as FilterAverage does. This means that if a data set is run
thorugh Add it gets the same number of rows as if it was run thorough
avarage.

In the manager code was fixed to differentiate between different raw
data files.
Also added some new filters and a new way of writing filtered data to
file.