A big design change in file combiner
We now take a enum which decides if we want to train a specific network
or some network combinations. This still have John code but it does not
run it, a future commit will remove it completely. We now run all
different training decisions through the same code and only differes at
the very end, right before we send it on to the factors