Filtering removing dead data and normalizing
The dead data should now be removed, we take away some more data points
thatn needed just to be sure. It now works so that we check if the
displacement is above the max movement, that is how we determine that it
should be removed. Then we remove all data points until the player
starts moving normaly again, not counting standing still as normal. This
should also help removing falling targets from the filtered data, which
might be a good thing.
Also made a new normalization filter which is proparly used by all
except displacement. We still need to figure out max displacement